Breaking change: config: Always require a default rule.
Previously, server provided an implicit “default deny” for any
request that did not match a rule. However, config files already
have semantics for expressing a default rule.

So it seems better to simply require all configs to express a
default rule, which puts the default policy firmly in the hands
of the service owner, and eliminates a weird request path.